New file created for macrocalculations
